feat: Upgrade to iced git version with IME support and refactor architecture
- Switch to iced git version for Linux IME support
- Refactor main.rs: move view, subscription, and event handling to state.rs
- Add context menu support for creating new collections in empty area
- Fix API compatibility for iced 0.13 (Space, svg, text_input::Id, etc.)
- Update all UI components to use new iced widget APIs
- Auto-focus and select text when creating new folders/collections

+++ src/ui/icons.rs
@@ -20,8 +20,8 @@ pub const SETTINGS_ICON: &str = r##"<svg viewBox="0 0 1024 1024" xmlns="http://w
 /// An SVG element that can be used in the UI
 pub fn loading_icon<'a, Message: 'a>(size: u16) -> Element<'a, Message> {
     svg(svg::Handle::from_memory(LOADING_ICON.as_bytes()))
-        .width(size)
-        .height(size)
+        .width(size as f32)
+        .height(size as f32)
         .into()
 }
 
@@ -34,7 +34,7 @@ pub fn loading_icon<'a, Message: 'a>(size: u16) -> Element<'a, Message> {
 /// An SVG element that can be used in the UI
 pub fn settings_icon<'a, Message: 'a>(size: u16) -> Element<'a, Message> {
     svg(svg::Handle::from_memory(SETTINGS_ICON.as_bytes()))
-        .width(size)
-        .height(size)
+        .width(size as f32)
+        .height(size as f32)
         .into()
 }
